Gemini-CLI-MCP/requirements.txt
mcp>=1.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0
asyncio
typing
logging
json5
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0